fix: add unitType for ingredient

+++ backend/module/ingredient/ingredientmodel/ingredient_update_price.go
@@ -3,7 +3,7 @@ package ingredientmodel
 import "backend/common"
 
 type IngredientUpdatePrice struct {
-	Price *float32 `json:"price" gorm:"column:price;"`
+	Price *int `json:"price" gorm:"column:price;"`
 }
 
 func (*IngredientUpdatePrice) TableName() string {
