+++ backend/module/stockreportdetail/stockreportdetailmodel/stock_report_detail_create.go
@@ -5,14 +5,14 @@ import (
 )
 
 type StockReportDetailCreate struct {
-	ReportId     string `json:"-" gorm:"column:reportId;"`
-	IngredientId string `json:"-" gorm:"column:ingredientId;"`
-	Initial      int    `json:"-" gorm:"column:initial;"`
-	Sell         int    `json:"-" gorm:"column:sell"`
-	Import       int    `json:"-" gorm:"column:import;"`
-	Export       int    `json:"-" gorm:"column:export;"`
-	Modify       int    `json:"-" gorm:"column:modify;"`
-	Final        int    `json:"-" gorm:"column:final;"`
+	ReportId     string  `json:"-" gorm:"column:reportId;"`
+	IngredientId string  `json:"-" gorm:"column:ingredientId;"`
+	Initial      float32 `json:"-" gorm:"column:initial;"`
+	Sell         float32 `json:"-" gorm:"column:sell"`
+	Import       float32 `json:"-" gorm:"column:import;"`
+	Export       float32 `json:"-" gorm:"column:export;"`
+	Modify       float32 `json:"-" gorm:"column:modify;"`
+	Final        float32 `json:"-" gorm:"column:final;"`
 }
 
 func (*StockReportDetailCreate) TableName() string {
